Added carousel

+++ SmartPromoCore.xcframework/ios-arm64_i386_x86_64-simulator/SmartPromoCore.framework/Headers/FSPDispatchQueue+SmartPromo.h
@@ -13,7 +13,7 @@ typedef void (^FSPBlock)(void);
 
 + (void) runInBackground: (dispatch_block_t) block;
 + (void) runInMain: (dispatch_block_t) block;
-+ (void) runInMainAfter: (int) delay block: (dispatch_block_t) block;
++ (void) runInMainAfter: (double) delay block: (dispatch_block_t) block;
 
 @end
